move candidate select files
